+++ redis/redis.go
@@ -46,6 +46,13 @@ func (c Client) Get(k string, v interface{}) (bool, error) {
 	return true, util.FromJSON([]byte(data), v)
 }
 
+// Delete deletes the stored value for the given key.
+// Deleting a non-existing key-value pair does NOT lead to an error.
+func (c Client) Delete(k string) error {
+	_, err := c.c.Del(k).Result()
+	return err
+}
+
 // Options are the options for the Redis client.
 type Options struct {
 	// Address of the Redis server, including the port.