+++ syncmap/map.go
@@ -35,6 +35,13 @@ func (m Store) Get(k string, v interface{}) (bool, error) {
 	return true, util.FromJSON(data.([]byte), v)
 }
 
+// Delete deletes the stored value for the given key.
+// Deleting a non-existing key-value pair does NOT lead to an error.
+func (m Store) Delete(k string) error {
+	m.m.Delete(k)
+	return nil
+}
+
 // NewStore creates a new Go sync.Map store.
 func NewStore() Store {
 	return Store{